+++ CHANGELOG.md
@@ -6,6 +6,13 @@
 1. [](#improved)
     * Make all form fields to extend `field.html.twig`
 
+# v2.16.4
+## mm/dd/2018
+
+1. [](#improved)
+    * Better handling of invalid file names during upload
+    * Better MIME type checking of files during file upload
+
 # v2.16.3
 ## 09/21/2018